refactor(bridge): extract whatsmeow logger adapter
